requirements.txt
astroid==1.5.3
coverage==4.4.2
ddt==1.1.1
django==1.11.8
django-countries==5.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==1.9.8
django-filter==1.1.0
django-localflavor==1.6.1
django-money==0.12.1
django-redis==4.8.0
django-simple-email-confirmation==0.22
django-vies==3.4.1
djangorestframework==3.7.3
djangorestframework-csv==2.0.0
docker-compose>=1.7.1,<2.0.0
dodgy==0.1.9
factory-boy==2.9.2
faker==0.8.7
freezegun==0.3.8
gunicorn==19.7.1
honcho==1.0.1
huey==1.5.5
ipdb==0.10.1
ipython==6.2.1
isort==4.2.15
mccabe==0.6.1
pep257==0.7.0
prospector==0.12.7
psycopg2==2.7.3.2
pycodestyle==2.0.0
pydocstyle==2.1.1
pyflakes==1.6.0
pylint==1.7.4
pylint-common==0.2.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytz==2017.3
refreshbooks==2.0
requests==2.18.4
responses==0.8.1
simplejson==3.13.2
Sphinx==1.6.5
virtualenv==15.1.0
Werkzeug==0.13
